// mileon-client/src/app/types/enum/terminalSettingsEnum.ts
export enum TerminalSettingsFieldsEnum {
  PangoUsername = 'שם משתמש פנגו',
  PangoPassword = 'סיסמה פנגו',
  CelloparkUsername = 'שם משתמש סלופארק',
  CelloparkPassword = 'סיסמה סלופארק',
  RefundInterfaceStatus = 'סטטוס ממשק החזרים',
  SolarParking = 'חניה סלולארית',
  SolarParkingSuspicion = 'השהיית ממשקי חניה סלולארית',
  RefundInterface = 'ממשק החזרים',
  PangoInquiry = 'תשאול פנגו',
  CelloparkInquiry = 'תשאול סלופארק',
}